pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cortex-memory"
version = "0.1.0"
description = "A local-first memory system for AI assistants that actually remembers."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Gerry Campbell", email = "gerry@phasefour.ai"}
]
keywords = ["ai", "memory", "assistant", "llm", "vector-search", "chromadb"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "chromadb>=0.4.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40.0"]
openai = ["openai>=1.0.0"]
all = ["anthropic>=0.40.0", "openai>=1.0.0"]

[project.scripts]
cortex = "cortex_memory.cli:main"

[project.urls]
Homepage = "https://github.com/gfcampbell/cortex-memory"
Repository = "https://github.com/gfcampbell/cortex-memory"
Issues = "https://github.com/gfcampbell/cortex-memory/issues"

[tool.setuptools.packages.find]
include = ["cortex_memory*"]
